[API] application_id
revisited commit 27ab674db1554d71009081151e0e24dd728598a8
make API more in line with window_added() and window_removed().
Also makes it possible to relate the callback to a specific window.

Small fix since GetApplicationID did not work yet.

+++ src/sawman.c
@@ -366,8 +366,7 @@ manager_call_handler( int           caller,
 
           case SWMCID_APPLICATION_ID_CHANGED:
                if (sawman->manager.callbacks.ApplicationIDChanged)
-                    *ret_val = sawman->manager.callbacks.ApplicationIDChanged( sawman->manager.context,
-                                                                      (unsigned long)call_ptr );
+                    *ret_val = sawman->manager.callbacks.ApplicationIDChanged( sawman->manager.context, call_ptr );
                break;
 
           default:
